pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docchat-jbegley"
version = "0.1.0"
description = "An AI agent for chatting with documents in your project"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "groq",
    "python-dotenv",
]

[project.scripts]
chat = "chat:main"

[tool.setuptools]
py-modules = ["chat"]
packages = ["tools"]
